fix: disable OIDC provider creation in EKS stand up
We do not make use of a OIDC provider in our current
configuration. With this parameter enabled, it adds to the
total start up time for creating an EKS cluster. It also
causes failures when Pulumi is executed on an AWS EC2
instance with restrictive roles like the LightSail's
defaults.