refactor: move .python-version file to the root, increase patch version (#144)
Most version managers interpret the `.python-version` file from the root and apply the
python version specified to all child folders.  Moving the `.pythong-version` file
to the root of the project ensures that the python version is the same for all
actions, but most importantly avoids a situation when running `./bin/setup_venv.sh`
from the root where the system python will be hooked for the venv.