fix: pip3 install now does not accept the `--use-feature=in-tree-build`
The feature flag `--use-feature=in-tree-build` was introduced in pip 21.1 and then
the feature was mainlined in pip 21.3.  With the flag in place `pip install` command
fails and `kic_utils` fails to build.

Based on the changelog which states that this behavior is default in pip 21.3+ and the current
build installs pip 22.1 we remove the flag.

fixes #145